requirement.txt
asttokens==2.4.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
findspark==2.0.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.1
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
Pygments==2.16.1
pyspark==3.5.0
pytest==7.4.3
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
six==1.16.0
spark==0.2.1
stack-data==0.6.3
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
wcwidth==0.2.9
